+++ Changes.md
@@ -14,6 +14,7 @@ Change History
 * Added paging of "/w ore" output.
 * Added a configuration option for the Watson command prefix.
 * Added a "/w file delete <prefix>" command to delete save files.
+* Added a "/w file expire <YYYY-MM-DD>" to delete files last modified before a specified date.
 
 1.4.7 (2013-01-23)
 ------------------